docker/sandbox/tools/t1/kook-notify/requirements.txt
# This file is @generated by PDM.
# Please do not edit it manually.

aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
apscheduler==3.11.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1; platform_python_implementation == "CPython" and sys_platform == "win32"
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6; platform_system == "Windows"
dify-plugin==0.0.1b77
dpkt==1.9.8
flask==3.0.3
frozenlist==1.6.0
gevent==24.11.1
greenlet==3.2.0; platform_python_implementation == "CPython"
h11==0.14.0
httpcore==1.0.8
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.6
khl-py==0.3.17
markupsafe==3.0.2
multidict==6.4.3
propcache==0.3.1
pycparser==2.22; platform_python_implementation == "CPython" and sys_platform == "win32"
pycryptodomex==3.22.0
pydantic==2.11.3
pydantic-core==2.33.1
pydantic-settings==2.9.1
pydub==0.25.1
python-dotenv==1.1.0
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
setuptools==79.0.0
sniffio==1.3.1
socksio==1.0.0
tiktoken==0.8.0
typing-extensions==4.13.2
typing-inspection==0.4.0
tzdata==2025.2; platform_system == "Windows"
tzlocal==5.3.1
urllib3==2.4.0
werkzeug==3.0.6
yarl==1.20.0
zope-event==5.0
zope-interface==7.2
